Refactored components to improve performance
